wallet: move methods to deprecated.go
This is a cleanup commit to move all deprecated methods into
`deprecated.go` - we still keep them to make sure the benchmark tests
can be run.
